# gui/src/assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
    "title": "FastAPI",
    "version": "0.1.0"
    },
    "paths": {
...
    "/api/v1/roles/": {
    "post": {
    "tags": [
    "Roles",
    "RBAC"
    ],
    "summary": "Rolecontroller.Create Role",
    "operationId": "RoleController_create_role_api_v1_roles__post",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/RoleCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/RoleRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Roles",
    "RBAC"
    ],
    "summary": "Rolecontroller.List Roles",
    "operationId": "RoleController_list_roles_api_v1_roles__get",
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_RoleRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/roles/{role_id}": {
    "delete": {
    "tags": [
    "Roles",
    "RBAC"
    ],
    "summary": "Rolecontroller.Delete Role",
    "operationId": "RoleController_delete_role_api_v1_roles__role_id__delete",
    "parameters": [
    {
    "name": "role_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Role Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Roles",
    "RBAC"
    ],
    "summary": "Rolecontroller.Get Role",
    "operationId": "RoleController_get_role_api_v1_roles__role_id__get",
    "parameters": [
    {
    "name": "role_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Role Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/RoleRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Roles",
    "RBAC"
    ],
    "summary": "Rolecontroller.Patch Role",
    "operationId": "RoleController_patch_role_api_v1_roles__role_id__patch",
    "parameters": [
    {
    "name": "role_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Role Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/RoleUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/RoleRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/permissions/": {
    "post": {
    "tags": [
    "Permissions",
    "RBAC"
    ],
    "summary": "Permissioncontroller.Create Permission",
    "operationId": "PermissionController_create_permission_api_v1_permissions__post",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PermissionCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PermissionRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Permissions",
    "RBAC"
    ],
    "summary": "Permissioncontroller.List Permissions",
    "operationId": "PermissionController_list_permissions_api_v1_permissions__get",
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_PermissionRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/permissions/{permission_id}": {
    "delete": {
    "tags": [
    "Permissions",
    "RBAC"
    ],
    "summary": "Permissioncontroller.Delete Permission",
    "operationId": "PermissionController_delete_permission_api_v1_permissions__permission_id__delete",
    "parameters": [
    {
    "name": "permission_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Permission Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Permissions",
    "RBAC"
    ],
    "summary": "Permissioncontroller.Get Permission",
    "operationId": "PermissionController_get_permission_api_v1_permissions__permission_id__get",
    "parameters": [
    {
    "name": "permission_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Permission Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PermissionRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Permissions",
    "RBAC"
    ],
    "summary": "Permissioncontroller.Patch Permission",
    "operationId": "PermissionController_patch_permission_api_v1_permissions__permission_id__patch",
    "parameters": [
    {
    "name": "permission_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Permission Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PermissionUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PermissionRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
...
    "HTTPValidationError": {
    "properties": {
    "detail": {
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    },
    "type": "array",
    "title": "Detail"
    }
    },
    "type": "object",
    "title": "HTTPValidationError"
    },
...
    "Page_PermissionRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/PermissionRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[PermissionRead]"
    },
    "Page_RoleRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/RoleRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[RoleRead]"
    },
...
    "PermissionAction": {
    "type": "string",
    "enum": [
    "c",
    "r",
    "u",
    "d"
    ],
    "title": "PermissionAction"
    },
    "PermissionCreate": {
    "properties": {
    "resource": {
    "type": "string",
    "title": "Resource"
    },
    "action": {
    "anyOf": [
    {
    "$ref": "#/components/schemas/PermissionAction"
    },
    {
    "type": "string"
    }
    ],
    "title": "Action"
    }
    },
    "type": "object",
    "required": [
    "resource",
    "action"
    ],
    "title": "PermissionCreate"
    },
    "PermissionRead": {
    "properties": {
    "permission_id": {
    "type": "integer",
    "title": "Permission Id"
    },
    "resource": {
    "type": "string",
    "title": "Resource"
    },
    "action": {
    "anyOf": [
    {
    "$ref": "#/components/schemas/PermissionAction"
    },
    {
    "type": "string"
    }
    ],
    "title": "Action"
    }
    },
    "type": "object",
    "required": [
    "permission_id",
    "resource",
    "action"
    ],
    "title": "PermissionRead"
    },
    "PermissionUpdate": {
    "properties": {
    "resource": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Resource"
    },
    "action": {
    "anyOf": [
    {
    "$ref": "#/components/schemas/PermissionAction"
    },
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Action"
    }
    },
    "type": "object",
    "title": "PermissionUpdate"
    },
    "RoleCreate": {
    "properties": {
    "name": {
    "type": "string",
    "title": "Name"
    },
    "permissions": {
    "anyOf": [
    {
    "items": {
    "$ref": "#/components/schemas/PermissionCreate"
    },
    "type": "array"
    },
    {
    "type": "null"
    }
    ],
    "title": "Permissions"
    }
    },
    "type": "object",
    "required": [
    "name"
    ],
    "title": "RoleCreate"
    },
    "RoleRead": {
    "properties": {
    "role_id": {
    "type": "integer",
    "title": "Role Id"
    },
    "name": {
    "type": "string",
    "title": "Name"
    },
    "permissions": {
    "items": {
    "$ref": "#/components/schemas/PermissionRead"
    },
    "type": "array",
    "title": "Permissions",
    "default": []
    }
    },
    "type": "object",
    "required": [
    "role_id",
    "name"
    ],
    "title": "RoleRead"
    },
    "RoleUpdate": {
    "properties": {
    "name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Name"
    }
    },
    "type": "object",
    "title": "RoleUpdate"
    },
...
    "ValidationError": {
    "properties": {
    "loc": {
    "items": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "integer"
    }
    ]
    },
    "type": "array",
    "title": "Location"
    },
    "msg": {
    "type": "string",
    "title": "Message"
    },
    "type": {
    "type": "string",
    "title": "Error Type"
    }
    },
    "type": "object",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "title": "ValidationError"
    }
    },
    "securitySchemes": {
    "APIKeyCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "med_voice_auth_token"
    }
    }
    }
    }